Added a submit post page and linked to it in the header
Also compared meteor packages with the course latest github repo and made certain changes to my own package to match the latest course app more closely.
So some of my code etc may or may not be identical to the code in the book. CORRCTED AN ERROR IN PREVIOUS COMMIT.